requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-ninja==1.1.0
flatbuffers==23.5.26
fonttools==4.47.2
humanfriendly==10.0
idna==3.6
imageio==2.33.1
imutils==0.5.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
onnxruntime==1.16.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.25.2
pydantic==2.5.3
pydantic_core==2.14.6
PyMatting==1.1.12
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
referencing==0.32.1
rembg==2.0.53
requests==2.31.0
rpds-py==0.16.2
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tifffile==2023.12.9
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
